requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.3
Werkzeug==3.1.5
vaderSentiment==3.3.2
